Change Microsoft.DotNet.ApiSymbolExtensions project type to Class library. Added Microsoft.DotNet.ApiSymbolExtensions.Tests test project
