+++ src/Microsoft.DotNet.ApiSymbolExtensions/Extensions/SymbolExtensions.cs
@@ -7,11 +7,11 @@ using System.Collections.Generic;
 
 namespace Microsoft.DotNet.ApiSymbolExtensions.Extensions
 {
-    internal static class SymbolExtensions
+    public static class SymbolExtensions
     {
-        private static SymbolDisplayFormat Format { get; } = GetSymbolDisplayFormat();
+        public static SymbolDisplayFormat Format { get; } = GetSymbolDisplayFormat();
 
-        private static SymbolDisplayFormat GetSymbolDisplayFormat()
+        public static SymbolDisplayFormat GetSymbolDisplayFormat()
         {
             // This is the default format for symbol.ToDisplayString;
             SymbolDisplayFormat format = SymbolDisplayFormat.CSharpErrorMessageFormat;
@@ -29,12 +29,12 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Extensions
             return format.WithParameterOptions(format.ParameterOptions & ~SymbolDisplayParameterOptions.IncludeParamsRefOut);
         }
 
-        internal static string ToComparisonDisplayString(this ISymbol symbol) =>
+        public static string ToComparisonDisplayString(this ISymbol symbol) =>
             symbol.ToDisplayString(Format)
                   .Replace("System.IntPtr", "nint") // Treat IntPtr and nint as the same
                   .Replace("System.UIntPtr", "nuint"); // Treat UIntPtr and nuint as the same
 
-        internal static IEnumerable<ITypeSymbol> GetAllBaseTypes(this ITypeSymbol type)
+        public static IEnumerable<ITypeSymbol> GetAllBaseTypes(this ITypeSymbol type)
         {
             if (type.TypeKind == TypeKind.Interface)
             {
@@ -53,7 +53,7 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Extensions
             }
         }
 
-        internal static bool IsEffectivelySealed(this ITypeSymbol type, bool includeInternals) =>
+        public static bool IsEffectivelySealed(this ITypeSymbol type, bool includeInternals) =>
             type.IsSealed || !HasVisibleConstructor(type, includeInternals);
 
         private static bool HasVisibleConstructor(ITypeSymbol type, bool includeInternals)
@@ -70,7 +70,7 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Extensions
             return false;
         }
 
-        internal static IEnumerable<ITypeSymbol> GetAllBaseInterfaces(this ITypeSymbol type)
+        public static IEnumerable<ITypeSymbol> GetAllBaseInterfaces(this ITypeSymbol type)
         {
             foreach (ITypeSymbol @interface in type.Interfaces)
             {
@@ -84,7 +84,7 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Extensions
                     yield return baseInterface;
         }
 
-        internal static bool IsVisibleOutsideOfAssembly(this ISymbol symbol, bool includeInternals, bool includeEffectivelyPrivateSymbols = false) =>
+        public static bool IsVisibleOutsideOfAssembly(this ISymbol symbol, bool includeInternals, bool includeEffectivelyPrivateSymbols = false) =>
             symbol.DeclaredAccessibility switch
             {
                 Accessibility.Public => true,
@@ -94,7 +94,7 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Extensions
                 _ => includeInternals && symbol.DeclaredAccessibility != Accessibility.Private,
             };
 
-        internal static bool IsEventAdderOrRemover(this IMethodSymbol method) =>
+        public static bool IsEventAdderOrRemover(this IMethodSymbol method) =>
             method.MethodKind == MethodKind.EventAdd ||
             method.MethodKind == MethodKind.EventRemove ||
             method.Name.StartsWith("add_", StringComparison.Ordinal) ||
