+++ src/Microsoft.DotNet.ApiSymbolExtensions/IAssemblySymbolLoader.cs
@@ -2,9 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.IO;
 
 namespace Microsoft.DotNet.ApiSymbolExtensions
